feat(tracer): add esmodule support
