--- packages/tracer/src/config/index.ts
@@ -1,2 +1,0 @@
-export * from './ConfigServiceInterface';
-export * from './EnvironmentVariablesService';
