+++ packages/tracer/src/index.ts
@@ -1,3 +1,1 @@
-export * from './Tracer';
-export * from './TracerInterface';
-export * from './middleware/middy';
+export { Tracer } from './Tracer.js';
