--- packages/tracer/src/provider/index.ts
@@ -1,2 +1,0 @@
-export * from './ProviderService';
-export * from './ProviderServiceInterface';
