+++ packages/tracer/src/types/Tracer.ts
@@ -1,10 +1,6 @@
-import type { ConfigServiceInterface } from '../config';
-import type { Handler } from 'aws-lambda';
-import type {
-  AsyncHandler,
-  LambdaInterface,
-  SyncHandler,
-} from '@aws-lambda-powertools/commons/types';
+import type { ConfigServiceInterface } from './ConfigServiceInterface.js';
+import type { HandlerMethodDecorator } from '@aws-lambda-powertools/commons/types';
+import type { Segment, Subsegment } from 'aws-xray-sdk-core';
 
 /**
  * Options for the tracer class to be used during initialization.
@@ -100,28 +96,51 @@ type CaptureMethodOptions = {
   captureResponse?: boolean;
 };
 
-type HandlerMethodDecorator = (
-  target: LambdaInterface,
+// eslint-disable-next-line  @typescript-eslint/no-explicit-any
+type AnyClassMethod = (...args: any[]) => any;
+// eslint-disable-next-line  @typescript-eslint/no-explicit-any
+type AnyClass = new (...args: any[]) => any;
+
+type MethodDecorator<T extends AnyClass> = (
+  target: InstanceType<T>,
   propertyKey: string | symbol,
-  descriptor:
-    | TypedPropertyDescriptor<SyncHandler<Handler>>
-    | TypedPropertyDescriptor<AsyncHandler<Handler>>
+  descriptor: TypedPropertyDescriptor<AnyClassMethod>
 ) => void;
 
-// TODO: Revisit type below & make it more specific
-type MethodDecorator = (
-  // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-  target: any,
-  propertyKey: string | symbol,
-  // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-  descriptor: TypedPropertyDescriptor<any>
-  // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-) => any;
+interface TracerInterface {
+  addErrorAsMetadata(error: Error, remote?: boolean): void;
+  addResponseAsMetadata(data?: unknown, methodName?: string): void;
+  addServiceNameAnnotation(): void;
+  annotateColdStart(): void;
+  captureAWS<T>(aws: T): void | T;
+  captureAWSv3Client<T>(service: T): void | T;
+  captureAWSClient<T>(service: T): void | T;
+  captureLambdaHandler(
+    options?: CaptureLambdaHandlerOptions
+  ): HandlerMethodDecorator;
+  captureMethod<T extends AnyClass>(
+    options?: CaptureMethodOptions
+  ): MethodDecorator<T>;
+  getSegment(): Segment | Subsegment | undefined;
+  getRootXrayTraceId(): string | undefined;
+  isTraceSampled(): boolean;
+  isTracingEnabled(): boolean;
+  putAnnotation: (key: string, value: string | number | boolean) => void;
+  putMetadata: (
+    key: string,
+    value: unknown,
+    namespace?: string | undefined
+  ) => void;
+  setSegment(segment: Segment | Subsegment): void;
+}
 
 export {
   TracerOptions,
   CaptureLambdaHandlerOptions,
   CaptureMethodOptions,
   HandlerMethodDecorator,
+  AnyClass,
+  AnyClassMethod,
   MethodDecorator,
+  TracerInterface,
 };
