+++ packages/tracer/src/types/index.ts
@@ -1,1 +1,10 @@
-export * from './Tracer';
+export {
+  TracerOptions,
+  CaptureLambdaHandlerOptions,
+  CaptureMethodOptions,
+  HandlerMethodDecorator,
+  AnyClass,
+  AnyClassMethod,
+  MethodDecorator,
+  TracerInterface,
+} from './Tracer.js';
