+++ packages/tracer/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "outDir": "./lib",
+        "outDir": "./lib/cjs",
         "rootDir": "./src",
     },
     "include": [
